requirements.txt
networkx
requests
praw
python-dotenv
logging